# perf: Optimize SourceLink creation, track state assignment