--- a/Core/include/Acts/EventData/MultiTrajectoryBackendConcept.hpp
+++ b/Core/include/Acts/EventData/MultiTrajectoryBackendConcept.hpp
@@ -130,7 +130,7 @@ concept MutableMultiTrajectoryBackend = CommonMultiTrajectoryBackend<T> &&
 
   {v.allocateCalibrated_impl(istate, dim)};
 
-  {v.setUncalibratedSourceLink_impl(istate, sl)};
+  {v.setUncalibratedSourceLink_impl(istate, std::move(sl))};
 
   {v.setReferenceSurface_impl(istate, surface)};