--- a/Core/include/Acts/EventData/TrackStateProxy.hpp
+++ b/Core/include/Acts/EventData/TrackStateProxy.hpp
@@ -676,10 +676,19 @@ class TrackStateProxy {
   /// @return The uncalibrated measurement source link
   SourceLink getUncalibratedSourceLink() const;
 
+  // This function will move to an rvalue reference in the next major version
   /// Set an uncalibrated source link
   /// @param sourceLink The uncalibrated source link to set
-  void setUncalibratedSourceLink(SourceLink&& sourceLink) requires(!ReadOnly) {
-    m_traj->setUncalibratedSourceLink(m_istate, std::move(sourceLink));
+  template<typename source_link_t>
+  void setUncalibratedSourceLink(source_link_t&& sourceLink) requires(!ReadOnly) {
+    m_traj->setUncalibratedSourceLink(m_istate, std::forward<source_link_t>(sourceLink));
+  }
+
+  /// Set an uncalibrated source link
+  /// @param sourceLink The uncalibrated source link to set
+  /// @deprecated Use the overload with an rvalue reference
+  void setUncalibratedSourceLink(const SourceLink& sourceLink) requires(!ReadOnly) {
+    m_traj->setUncalibratedSourceLink(m_istate, SourceLink{sourceLink});
   }
 
   /// Check if the point has an associated uncalibrated measurement.