This PR adds a benchmark for some `SourceLink` operations, and includes some optimization of creation and assignment into track states.

- Benchmark `main` 
```
Creating source link
  20000 runs of 100000 iteration(s), 7248.0ms total, 357.7920+/-0.1250µs per run, 3.578+/-0.395ns per iteration
Copy construct source link
20000 runs of 100000 iteration(s), 2583.0ms total, 126.8750+/-0.0349µs per run, 1.269+/-0.110ns per iteration
Copy then move construct source link
20000 runs of 100000 iteration(s), 4265.5ms total, 210.3750+/-0.0655µs per run, 2.104+/-0.207ns per iteration
Optional assignment
20000 runs of 100000 iteration(s), 9581.2ms total, 475.5420+/-0.1245µs per run, 4.755+/-0.394ns per iteration
Create track state
20000 runs of 10000 iteration(s), 12651.4ms total, 174.1250+/-0.4740µs per run, 17.413+/-4.740ns per iteration
Assign source link to track state
20000 runs of 10000 iteration(s), 6752.2ms total, 275.4160+/-0.6607µs per run, 27.542+/-6.607ns per iteration
```

- Benchmark `feat`
```
Creating source link
  20000 runs of 100000 iteration(s), 1528.3ms total, 75.2500+/-0.0118µs per run, 0.752+/-0.037ns per iteration
Copy construct source link
20000 runs of 100000 iteration(s), 2568.1ms total, 126.6250+/-0.0190µs per run, 1.266+/-0.060ns per iteration
Copy then move construct source link
20000 runs of 100000 iteration(s), 4225.4ms total, 209.7500+/-0.0177µs per run, 2.098+/-0.056ns per iteration
Optional assignment
20000 runs of 100000 iteration(s), 1827.6ms total, 90.5830+/-0.0027µs per run, 0.906+/-0.009ns per iteration
Create track state
20000 runs of 10000 iteration(s), 12539.9ms total, 177.1670+/-0.5062µs per run, 17.717+/-5.062ns per iteration
Assign source link to track state
20000 runs of 10000 iteration(s), 4855.1ms total, 189.8750+/-0.6325µs per run, 18.988+/-6.325ns per iteration
```

(measurements on my laptop)

- Raw creation time goes from about 3.6ns to about 0.8ns (about a 5x speedup)
- Move assignment of an empty optional from a source link goes from 4.8ns to 0.9ns (including a raw source link construction of of 0.8ns)
- Assignment of a source link into a track state (which holds an empty optional) goes from 27.5ns to 19ns (40% speedup), which includes track state creation (17.7ns which is unchanged by the optimizations in this PR).