@goetzgaycken 2024-08-01
why is creation now faster than copy ?