@paulgessinger 2024-08-01
@goetzgaycken I think it's because the creation needs to do less. The construction into the local buffer is known at compile time, so no pointer dereferencing is needed at all. Any type of copy or move operation needs to consult the type-specific handler pointer to figure out what to do, even if it's a trivial memcpy in the end.