@paulgessinger 2024-08-06
Sonar is happy now, and I've pushed an update which should make the Athena build happy.

I'll add another PR to make the API breaking changes in v37.